+++ src/style.ts
@@ -1,0 +1,4 @@
+import OlStyle from 'ol/style/Style.js';
+
+/** Single instance of an ol empty style */
+export const EmptyStyle = new OlStyle();